🚚 Renamed application_command to app_command
